% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaners.R
\name{clean_real}
\alias{clean_real}
\title{Clean real datasets}
\usage{
clean_real(dirty_d, target_nt, min_nt, trim_top, trim_bottom)
}
\description{
Removes selected x% of top and low trials and equalizes the number of trials by resampling trials until the desired number is achieved
}
