% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{compute_agg}
\alias{compute_agg}
\title{Compute aggregate measures}
\usage{
compute_agg(d)
}
\arguments{
\item{d}{a dataframe in which we will compute the aggregate measures}
}
\description{
This function computes the mean and median values of all the variables in a dataframe grouped by the variable
\emph{iter}, \emph{id} and \emph{condition}
}
