% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{compute_agg_dt}
\alias{compute_agg_dt}
\title{Compute aggregate measures using data.table
Tidy can manage data.tables so no need to convert back}
\usage{
compute_agg_dt(d)
}
\description{
Compute aggregate measures using data.table
Tidy can manage data.tables so no need to convert back
}
