% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{compute_one_iteration}
\alias{compute_one_iteration}
\title{Computes one iteration of the simulation}
\usage{
compute_one_iteration(d, ni, np, ntc, ids)
}
\arguments{
\item{d}{full dataset}

\item{ni}{the number of iterations to be computed at once}

\item{np}{the number of participants to be used in the iteration}

\item{ntc}{number of trials per condition}

\item{ids}{unique ids of each participant in the dataset}
}
\description{
Computes one iteration of the simulation
}
