% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{compute_pp}
\alias{compute_pp}
\title{Compute preprocessing pipelines}
\usage{
compute_pp(d)
}
\arguments{
\item{d}{a dataframe in which we will compute the different pipelines}
}
\description{
This function computes the designed preprocessing pipelines described in the manuscript and adds them as new variables
}
\details{
The dataframe must contain the following variables:
\describe{
  \item{iter}{# of the iteration}
  \item{id}{id of the participant}
  \item{rt_raw}{Reaction time}
  \item{rt_log}{Log-transform of the reaction time}
  \item{rt_inv}{Inverse of the reaction time}
}
}
