% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{compute_test}
\alias{compute_test}
\title{Compute t-test}
\usage{
compute_test(d)
}
\description{
This function computes the t-test values of all the variables in a dataframe grouped by the variables
\emph{iter}
}
\details{
It computes the t-test by calculating the difference between condition per participant and then computing a one-sample t-test.

#' @param d a dataframe in which we will compute the tests
}
