% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaners.R
\name{format_flexicon}
\alias{format_flexicon}
\title{Format flexicon data}
\usage{
format_flexicon(in_file, out_file = NULL)
}
\description{
It changes names and selects and renames the relevant columns, id and rt_raw
}
