% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaners.R
\name{format_stroop}
\alias{format_stroop}
\title{Format Stroop data}
\usage{
format_stroop(in_file, out_file = NULL)
}
\description{
It changes names and selects and renames the relevant columns, id and rt_raw
}
