% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaners.R
\name{format_synth}
\alias{format_synth}
\title{Format the simulated data csv's originally generated using python}
\usage{
format_synth(in_file, out_file = NULL)
}
\description{
It changes names and selects and renames the relevant columns, id and rt_raw
}
