% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_generator.R
\name{gen_dummy}
\alias{gen_dummy}
\title{Generate dummy datasets for testing}
\usage{
gen_dummy(ni = 4, np = 3, nipi = 2)
}
\description{
The naming makes reference to the function in which they must be used, for example dummy_agg can be used as a parameter for compute_agg
}
