% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_generator.R
\name{gen_dummy_test}
\alias{gen_dummy_test}
\title{Generate a dataset to be used as parameter in compute_test}
\usage{
gen_dummy_test(ni = 4, np = 3, nipi = 2)
}
\description{
Generate a dataset to be used as parameter in compute_test
}
