% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_ids}
\alias{get_ids}
\title{Get unique ids of the dataset}
\usage{
get_ids(d)
}
\arguments{
\item{d}{a dataset}
}
\description{
It will return a vector with the unique participant id in the dataset
}
