% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_ntc}
\alias{get_ntc}
\title{Get the number of trials per condition}
\usage{
get_ntc(d)
}
\arguments{
\item{d}{a dataset}
}
\description{
This function returns the number of trial per condition
}
\details{
It also asserts that the number of trial is equal for all participants, and that the number of trials is even for all participants.
}
