% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Get file and launch iterations}
\usage{
run_simulation(file, ni, np, nipi)
}
\arguments{
\item{file}{path to the file containing the data over which we will run the simulations}

\item{ni}{number of iterations that will be performed}

\item{np}{number of participants to be included in each of the simulation iterations}

\item{nipi}{number of iterations per iteration}
}
\description{
The main function to be called when invoking the simulation.
}
\details{
The difference between \emph{ni} and \emph{nipi} is that \emph{ni} is the total number of iterations to be run during the simulation,
the relevant value, while \emph{nipi} has computational purposes as we will run iterations in batches of size \emph{nipi}
this way we speed computations a bit. If confusing it can be set to 1, slowing things down, but only one iteration will be run per
loop making the process easier to understand.
}
