% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sample_data}
\alias{sample_data}
\title{Sample data}
\usage{
sample_data(d, ni, np, ids, replace_id = TRUE)
}
\arguments{
\item{d}{full dataset}

\item{ni}{the number of iterations we will perform in one chunk}

\item{np}{number of participants to sample}

\item{ids}{a vector with the ids of each of participant}

\item{replace_id}{flag that indicates if participants ids will be replaced (1:np) or not}
}
\description{
This function generates \emph{ni} samples of \emph{np} participants in each sample.
To avoid the case in which the same participant is sampled more than once in the same sample, and then all repetitions will be
grouped together when aggregated, the ids of the participants are replaced by numbers from 1 to np in each of the samples if
the \emph{replace_id} flag is set to true.
}
\details{
We give the option of including several samples (i.e. iterations) in a single dataframe to speed up the computations to compute several
iterations at once.
}
