% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{t_test}
\alias{t_test}
\title{t-test lite version}
\usage{
t_test(x, np)
}
\arguments{
\item{x}{vector of values}

\item{np}{number of observations (df=np-1)}
}
\description{
It computes a one sample t-test but \emph{quicker} than the t.test function because it does less things.
}
