% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{shuffle_cond}
\alias{shuffle_cond}
\title{Shuffle condition labels by iteration and parcitipant}
\usage{
shuffle_cond(d)
}
\description{
Recieves a dataframe with at least the columns id, iter and condition
}
